+++ CHANGELOG.md
@@ -6,13 +6,13 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
-- Orphaned resources are deleted ([#xxx])
+- Orphaned resources are deleted ([#249])
 
 ### Changed
 
-- `operator-rs` `0.24.0` -> `0.25.2` ([#xxx])
+- `operator-rs` `0.24.0` -> `0.25.2` ([#249])
  
-[#xxx]: https://github.com/stackabletech/hdfs-operator/pull/xxx
+[#249]: https://github.com/stackabletech/hdfs-operator/pull/249
 
 ## [0.5.0] - 2022-09-06