+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- Orphaned resources are deleted ([#249])
+
+### Changed
+
+- `operator-rs` `0.24.0` -> `0.25.2` ([#249])
+
+[#249]: https://github.com/stackabletech/hdfs-operator/pull/249
+
 ## [0.5.0] - 2022-09-06
 
 ### Changed